pyproject.toml
[project]
name = "slidesynth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.12",
    "docling>=2.72.0",
    "fastapi>=0.128.5",
    "jinja2>=3.1.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=13.0",
    "playwright>=1.58.0",
    "langchain-groq>=1.1.2",
]

[project.optional-dependencies]
# Enable LangExtract for source-grounded semantic extraction from parsed papers.
# Requires a Google AI Studio API key (LANGEXTRACT_API_KEY) and
# config: parsing.enhanced_extraction: true
enhanced = [
    "langextract>=0.2.0",
]
# Run every agent role on a local Ollama server — free, no API keys needed.
# Requires Ollama installed and running: https://ollama.com
local = [
    "langchain-ollama>=0.3.0",
]
